# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenCut Video Editor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  /** Duration in seconds */
  duration: number;
  /** Frames per second */
  fps: number;
  /** Resolution string e.g. 1920x1080 */
  resolution: string;
  /**
     * JSON-encoded timeline state
     * @nullable
     */
  timeline?: string | null;
  assetCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  fps?: number;
  resolution?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  thumbnail?: string;
  duration?: number;
  fps?: number;
  resolution?: string;
  timeline?: string;
}

export interface Asset {
  id: number;
  projectId: number;
  name: string;
  /** video, audio, image */
  type: string;
  url: string;
  /** File size in bytes */
  size: number;
  /**
     * Duration in seconds for video/audio
     * @nullable
     */
  duration?: number | null;
  /** @nullable */
  thumbnail?: string | null;
  createdAt: string;
}

export interface AssetInput {
  name: string;
  type: string;
  url: string;
  size: number;
  duration?: number;
  thumbnail?: string;
}

export interface ProjectStats {
  totalProjects: number;
  totalAssets: number;
  totalDuration: number;
  recentCount: number;
}