# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OpenCut Video Editor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Video project operations
  - name: assets
    description: Media asset operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/assets:
    get:
      operationId: listProjectAssets
      tags: [assets]
      summary: List media assets for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Add a media asset to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"

  /projects/{id}/assets/{assetId}:
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Remove an asset from a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get aggregate stats across all projects
      responses:
        "200":
          description: Project statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  /projects/recent:
    get:
      operationId: listRecentProjects
      tags: [projects]
      summary: Get recently modified projects
      responses:
        "200":
          description: Recent projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      required: [id, name, createdAt, updatedAt, duration, fps, resolution, assetCount]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        duration:
          type: number
          description: Duration in seconds
        fps:
          type: integer
          description: Frames per second
        resolution:
          type: string
          description: "Resolution string e.g. 1920x1080"
        timeline:
          type: ["string", "null"]
          description: JSON-encoded timeline state
        assetCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        fps:
          type: integer
          default: 30
        resolution:
          type: string
          default: "1920x1080"

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        thumbnail:
          type: string
        duration:
          type: number
        fps:
          type: integer
        resolution:
          type: string
        timeline:
          type: string

    Asset:
      type: object
      required: [id, projectId, name, type, url, size, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "video, audio, image"
        url:
          type: string
        size:
          type: integer
          description: File size in bytes
        duration:
          type: ["number", "null"]
          description: Duration in seconds for video/audio
        thumbnail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    AssetInput:
      type: object
      required: [name, type, url, size]
      properties:
        name:
          type: string
        type:
          type: string
        url:
          type: string
        size:
          type: integer
        duration:
          type: number
        thumbnail:
          type: string

    ProjectStats:
      type: object
      required: [totalProjects, totalAssets, totalDuration, recentCount]
      properties:
        totalProjects:
          type: integer
        totalAssets:
          type: integer
        totalDuration:
          type: number
        recentCount:
          type: integer