# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenCut Video Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  projectId: number;
  name: string;
  /** video, audio, image */
  type: string;
  url: string;
  /** File size in bytes */
  size: number;
  /**
     * Duration in seconds for video/audio
     * @nullable
     */
  duration?: number | null;
  /** @nullable */
  thumbnail?: string | null;
  createdAt: Date;
}