# lib/api-zod/src/generated/types/assetInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenCut Video Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface AssetInput {
  name: string;
  type: string;
  url: string;
  size: number;
  duration?: number;
  thumbnail?: string;
}