# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenCut Video Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  /** Duration in seconds */
  duration: number;
  /** Frames per second */
  fps: number;
  /** Resolution string e.g. 1920x1080 */
  resolution: string;
  /**
     * JSON-encoded timeline state
     * @nullable
     */
  timeline?: string | null;
  assetCount: number;
  createdAt: Date;
  updatedAt: Date;
}