# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenCut Video Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectStats {
  totalProjects: number;
  totalAssets: number;
  totalDuration: number;
  recentCount: number;
}