# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OpenCut Video Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  thumbnail?: string;
  duration?: number;
  fps?: number;
  resolution?: string;
  timeline?: string;
}